x86: boot: support minigzip bzImage compression
Android OTA system computes very efficient diffs of compressed files
if the deflate() algorithm it has access to is the same version as
used to create the original file. Here we add support for compressing
the kernel with 'minigzip' which uses the deflate() inside zlib.
This is much better than using 'gzip' as that tool has a very old
version of deflate() inside the gzip codebase instead of linking against
zlib.

Change-Id: I58be92d6a239814ad4510efe40f9f1e89b9f977c
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>
Reviewed-on: https://android.intel.com/169165
Reviewed-by: Uyyala, Sridhar <sridhar.uyyala@intel.com>
Reviewed-by: Gumbel, Matthew K <matthew.k.gumbel@intel.com>