REVERTME [VPG]: drm/i915: adding VLV specific sprite surface address register write.
The sprite surface address register is different for VLV.
The Macros to get the address of the VLV sprite address cannot be used
directly as there is no one on one mapping like HSW. For some reasons
there is no map stored in the kernel.
This patch has the mapping logic ported from MCG kernel.
This can be used as a temporary fix for the mapping problem.
This includes code to do plane id -2 in order to achive the mapping.If this
logic is made part of HWC then this patch can be modified to remove that logic
like it exists in MCG - HWC.

Issue: AIAF-3656
Change-Id: I7b9c75fa31f7bbb23f44fd288806764968806218
Signed-off-by: vanditak <vandita.kulkarni@intel.com>
Sugned-off-by: Gajanan Bhat <gajanan.bhat@intel.com>
Reviewed-on: https://android.intel.com/166276
Reviewed-by: Persson, Mikael <mikael.persson@intel.com>
Reviewed-by: Odeberg, Hans <hans.odeberg@intel.com>
Tested-by: Odeberg, Hans <hans.odeberg@intel.com>