IRDA-495: USB: mtp: fix the buffer overflow issue for transfers larger than 2GB
When the user connects the tablet to the host and tries to transfer a
file larger than 2 GB, the kernel instantly freezes and panics. This
patch fixes this issue.
The len variable is declared as int so when the length of file which is
larger than 2 GB is assigned to it, it becomes negative and invalid
length is assigned to read_req->length causing the buffer overwrite later.
To fix this we just make the len, the same type as count.

Bug: IRDA-495
Change-Id: I8897f6b47dd0a2b085b624de01ec330198f9c977
Signed-off-by: Manish Regmi <manish.regmi@intel.com>