MUST_REBASE [VPG]: adf/intel/vlv: sprite src rectangle fixed point support
The src rectangle of the sprite plane is in 16.16 fixed point format.
Convert to integer in the flip function.

MUST_REBASE: Since Google ADF framework is not in upstream and also
we cannot have two display drivers, other one being i915, we need to
work on i915 and ADF convergence path before it can be upstreamed

For: GMINL-3547
Change-Id: I5f32e2495bbd7399b2fbf6a5d3e45de74fa7f1e5
Signed-off-by: Pallavi <pallavi.g@intel.com>