OV5693: Emulate R_PWN (PWDNB) tied to DOVDD in SW
According to the OV5693 DS PWDND should be tied to DOVDD which is not
the case on ECS. By emulating this behavior in SW we get rid of the
"lost arbitration" errors that sometimes occur on ECS. Arbitration
problems should not occur on ECS I2C1 as there is only one master.
The theory is that OV5693 or AD5823 is misbehaving during startup,
temporarily jamming the I2C bus.

Change-Id: Idbb69a00cd054e870b92c8561238fdd0815cfe27
Tracked-On: GMINL-1773
Signed-off-by: Johan Mossberg <johan.mossberg@intel.com>
Reviewed-on: https://icggerrit.ir.intel.com/32385
Reviewed-by: Jing Tao <jing.tao@intel.com>
Approver: Wei Tao <wei.tao@intel.com>
Reviewed-by: Wei Tao <wei.tao@intel.com>