REVERTME: Fix racing timers in suspend process
BZ: GMINL-4192

We notice similar patch and discussion on https://lkml.org/lkml/2014/7/25/635
Different with this patch, that patch also take hibernation process into
account.

If kernel sovle the race in future, this patch can be reverted.

Detail of the issue:

During "all radio on suspend" test (google acceptance test), we observed
that sometimes the system wake immediately, just after it enter mwait.
The trace shows that the racing tick_sched_timer cause this wakeup.

The race happening in a couple of location is:

  1. disable IRQs   		(e.g. arch_suspend_disable_irqs())
     ...
  2. syscore_suspend
      -> timekeeping_suspend()
       -> clockevents_notify(SUSPEND)
        -> tick_suspend()   	(timers are turned off here)
     ...
  3. mwait            		(wait for wake-IRQ here)

Between steps 1 and 2 the timers can still generate interrupts that are
not handled and stay pending until step 3. That pending IRQ causes an
immediate - spurious - wake. In our trace, it's always tick_sched_timer.

The solution is to move the clockevents suspend/resume notification out
of the syscore_suspend step and explictly call them at the appropriate
time in the suspend paths. I.e. timers are suspend before IRQs get
disabled.

Change-Id: I11df90b1433653d768a46e076633798b5d5822c0
Signed-off-by: huchen2 <hu1.chen@intel.com>