MUST_REBASE [VPG]: drm/i915: Add wrapper function to get platform gen id
This wrapper function gets the gen id from which we can decide
whether it is a baytrail or cherrytrail device.

MUST_REBASE: Since Google ADF framework is not in upstream and also
we cannot have two display drivers, other one being i915, we need to
work on i915 and ADF convergence path before it can be upstreamed.

Issue: GMINL-5193
Change-Id: I87a8ba2eba06fa25ad5c16da1d45f08280cf5dc4
Signed-off-by: Deepak M <m.deepak@intel.com>