MUST_REBASE [VPG]: intel/core: Support for PIPE C
This patch adds support for PIPE C in the ADF driver.
The changes are:
1. Added simplified macros _PIPE2, _PIPE3 and _TRANSCODER2.
2. Added offfsets for pipe, transcoder, pallette and cursors, to
   support new macros.
3. Change the old register definitions to use _PIPE2 and _TRANSCODER2
   instead of _PIPE and _TRANSCODER. This will help to get right
   offsets for PIPEC.
4. Add CHV specific definitions in a new file.

MUST_REBASE: Since Google ADF framework is not in upstream and also
we cannot have two display drivers, other one being i915, we need to
work on i915 and ADF convergence path before it can be upstreamed.

Issue: GMINL-5544
Change-Id: I1d1455fd2a685126e3655f18dc06cfd3528b78fc
Signed-off-by: Shashank Sharma <shashank.sharma@intel.com>
Signed-off-by: Deepak M <m.deepak@intel.com>