MUST_REBASE [VPG]: video/adf/intel: Enable DSI PLL for both DSI0 and DSI1 for dual link
For Dual link MIPI Panels, dsipll clock for both DSI0 and DSI1 needs to be enabled.

MUST_REBASE: Since Google ADF framework is not in upstream and also
we cannot have two display drivers, other one being i915, we need to
work on i915 and ADF convergence path before it can be upstreamed.

For: GMINL-5276
Change-Id: I8ec00e595a21d5ff95cf28584801420b4669bc17
Signed-off-by: Gaurav K Singh <gaurav.k.singh@intel.com>