MUST_REBASE [VPG]: adf/intel/vlv: Brightness fix for edp
This patch fixes the issue of unable to change brightness in edp + dp
scenario.

MUST_REBASE: Since Google ADF framework is not in upstream and also
we cannot have two display drivers, other one being i915, we need to
work on i915 and ADF convergence path before it can be upstreamed.

Issue: GMINL-5485
Change-Id: Iaad6517b5101ef1d955feff71219287a34b86e7a
Signed-off-by: Thulasimani, Sivakumar <sivakumar.thulasimani@intel.com>