REVERTME [VPG]: drm/i915: DSI port initialization patch
For single link MIPI DSI panels on port C, port value for
OTP, enable and disable sequence (from VBT block 53) is set
to 0 and it doesnot match the MIPI DSI port to enable.
Due to this, MIPI on port C is blanking out.

It is expected that for single link MIPI DSI panels, port
value from VBT Block 2 should be used.

In this patch, if the MIPI DSI panel is single link, then
fix is made to use port value from Block 2 to decide on
which port the OTP, enable and disable packets has to be sent
instead of using port id field from OTP, enable and disable
sequence (from vbt Block 53).

This is tagged as revertme as a fix for this issue exists in upstream.

Upstream-Repo: http://anongit.freedesktop.org/git/drm-intel
Upstream-commit: f915084edc5aad351d6a4675b0bcdded9a00090d
commit f915084edc5aad351d6a4675b0bcdded9a00090d
Author: Gaurav K Singh <gaurav.k.singh@intel.com>
Date:   Wed Dec 10 22:07:40 2014 +0530

    drm/i915: Changes related to the sequence port no for

    From now on for both DSI Ports A & C, the seq_port value has been
    set to 0. seq_port value is parsed from Sequence block#53 of VBT.
    So, for packets that needs to be read/write for DSI single link on
    Port A and Port C will now be based on the DVO port from VBT block 2
    instead of seq_port.

    Signed-off-by: Gaurav K Singh <gaurav.k.singh@intel.com>
    Reviewed-by: Jani Nikula <jani.nikula@intel.com>
    Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>

Issue: IMINAN-19711
Change-Id: If10027adf23316b46939892e2c5813f11be3b77d
Signed-off-by: Tajunnisha N <tajunnisha.n@intel.com>
Signed-off-by: Prathik Chandrashekar <prathik.chandrashekar@intel.com>