dw_i2c: Fix hold and setup time issue
If the device configuration is updated by firmware and
the information is not passed to driver via firmware
interfaces like ACPI or DT parameters, then that information
will be lost once the device either resets or suspends. This
patch checks the boot value of hold and setup time during
device probe and updates the driver hold and setup time params.
This fix will ensure that the updates to hold and setup time
values are not lost after device suspend or resets.

Change-Id: I7c4a387e4a09e34b925fac66f717aa4c43d8184d
Signed-off-by: Sathyanarayanan Kuppuswamy <sathyanarayanan.kuppuswamy@intel.com>