iio: gyro: bmg160: remove redundant field
Replace the 'timestamp' field in struct bmg160_data with the identically
named field in iio_poll_func and with calls to iio_get_time_ns().

The reported timestamps may be slightly different, but the advantage is
that we no longer assume that the buffer of bmg160 is triggered by its
own trigger.

Change-Id: Ia8070b250954f10faddc8f1967f9dd0310b9ac4a
Signed-off-by: Vlad Dogaru <vlad.dogaru@intel.com>